Do not require requirements.txt to be present when upgrading the dependencies

+++ update_requirements.sh
@@ -5,7 +5,10 @@ cd "$(dirname -- "${BASH_SOURCE[0]}")"
 VENV_DIR=.venv
 
 rm -rf "$VENV_DIR"
-mv requirements.txt "requirements.txt.$(date "+%Y%m%d_%H%M%S").bak"
+if [[ -f requirements.txt ]]
+then
+    mv requirements.txt "requirements.txt.$(date "+%Y%m%d_%H%M%S").bak"
+fi
 
 python3 -m venv "$VENV_DIR"
 source "$VENV_DIR/bin/activate"
